Refactor event emitting from Android
Fix case where a push is received during app startup but the native module is not fully initialised.

In this case, the call to getJSModule from ReactContext fails.

The solution is to track initilisation state of the module.

Event emission is further refactored in general to reduce duplication by adding emit, maybeEmit, and jsQueuedEmit helpers.

The cold-start caching of deferred deep link data and opened push notification events is generified into a queue that is processed once the JS listeners are registered.