Fix in-app inbox item handling of null data payloads on iOS
The KSInAppInboxItem model defines data as _Nullable, but the mapping to NSDictionary for passing to RN through the bridge doens't handle this case properly.

It will try to pass nil to the dictionary literal and throw an exception as dictionaries are not allowed to contain simple types.

Fix by passing NSNull instead if the data is nil.